+++ lib/Dialect/TTNN/Utils/OptimizerOverrides.cpp
@@ -159,7 +159,7 @@ void OptimizerOverridesHandler::addInputLayoutOverride(
   inputLayoutOverrides[opName] = params;
 }
 void OptimizerOverridesHandler::addInputLayoutOverride(
-    StringRef opName, SmallVector<int64_t> operandIdxes) {
+    StringRef opName, SmallVector<int64_t> &operandIdxes) {
   inputLayoutOverrides[opName] =
       InputLayoutOverrideParams{std::move(operandIdxes)};
 }
@@ -168,7 +168,7 @@ void OptimizerOverridesHandler::addOutputLayoutOverride(
   outputLayoutOverrides[opName] = params;
 }
 void OptimizerOverridesHandler::addOutputLayoutOverride(
-    StringRef opName, SmallVector<int64_t> grid, BufferType bufferType,
+    StringRef opName, SmallVector<int64_t> &grid, BufferType bufferType,
     TensorMemoryLayout tensorMemoryLayout, tt::ttnn::Layout memoryLayout,
     tt::DataType dataType) {
   outputLayoutOverrides[opName] = OutputLayoutOverrideParams{
